Remove DOWNLOAD_SRPMS flag and functionality
This flag is broken in 2.0. When we query for the spec name and
version, the file can't be parsed due to new macros within 2.0
specs. These macros are resolved in the package mariner_macros
but since we are building outside of a chroot, this package can't
be loaded.

Remove this flag for now as there is not a strong need fgor it.
Future fixes will require a chroot so that mariner_macros is
present or loaded.