Kernel ARM64 configs parity with 2.0
For this commit, I analyzed all configs which were set to y/m in
the 5.15 kernel but not set or missing in 6.1. I determined whether
they were supported, refactored, or depricated in 6.1. In the first
two cases, I enable the missing config or its new equivalent.

This analysis was done by using a Mariner container, copying the
6.1 kernel source and running 'make menuconfig' until the missing
configs could all be explained via a commit in the kernel source.